Support shape constants in ONNX import and tests
Adds support for shape constants in ONNX models by updating constant node handling, code generation, and type conversion logic. Introduces a new test and model for shape constants, and updates binary operation error messages and gather node codegen. Also refines ONNX IR graph building to convert constant tensors to shape type when used in binary ops with shapes.

+++ crates/burn-import/src/burn/node/gather.rs
@@ -98,7 +98,7 @@ impl<PS: PrecisionSettings> NodeCodegen<PS> for GatherNode {
                             let output = &self.output.name();
                             quote! {
                                 let input_shape = &#input_shape_name;
-                                let #output = input_shape[#idx] as i64;
+                                let #output = input_shape[#idx];
                             }
                         }
                         _ => panic!(
